+++ src/nw_trace_diag_avx2_256_64.c
@@ -82,7 +82,7 @@ static inline int64_t _mm256_extract_epi64_rpl(__m256i a, int imm) {
 
 
 static inline void arr_store_si256(
-        int *array,
+        int8_t *array,
         __m256i vWH,
         int32_t i,
         int32_t s1Len,
@@ -90,16 +90,16 @@ static inline void arr_store_si256(
         int32_t s2Len)
 {
     if (0 <= i+0 && i+0 < s1Len && 0 <= j-0 && j-0 < s2Len) {
-        array[(i+0)*s2Len + (j-0)] = (int64_t)_mm256_extract_epi64_rpl(vWH, 3);
+        array[(i+0)*s2Len + (j-0)] = (int8_t)_mm256_extract_epi64_rpl(vWH, 3);
     }
     if (0 <= i+1 && i+1 < s1Len && 0 <= j-1 && j-1 < s2Len) {
-        array[(i+1)*s2Len + (j-1)] = (int64_t)_mm256_extract_epi64_rpl(vWH, 2);
+        array[(i+1)*s2Len + (j-1)] = (int8_t)_mm256_extract_epi64_rpl(vWH, 2);
     }
     if (0 <= i+2 && i+2 < s1Len && 0 <= j-2 && j-2 < s2Len) {
-        array[(i+2)*s2Len + (j-2)] = (int64_t)_mm256_extract_epi64_rpl(vWH, 1);
+        array[(i+2)*s2Len + (j-2)] = (int8_t)_mm256_extract_epi64_rpl(vWH, 1);
     }
     if (0 <= i+3 && i+3 < s1Len && 0 <= j-3 && j-3 < s2Len) {
-        array[(i+3)*s2Len + (j-3)] = (int64_t)_mm256_extract_epi64_rpl(vWH, 0);
+        array[(i+3)*s2Len + (j-3)] = (int8_t)_mm256_extract_epi64_rpl(vWH, 0);
     }
 }
 
@@ -122,7 +122,7 @@ parasail_result_t* FNAME(
     int64_t * const restrict s2 = s2B+PAD; /* will allow later for negative indices */
     int64_t * const restrict H_pr = _H_pr+PAD;
     int64_t * const restrict F_pr = _F_pr+PAD;
-    parasail_result_t *result = parasail_result_new_trace(s1Len, s2Len, 32, sizeof(int));
+    parasail_result_t *result = parasail_result_new_trace(s1Len, s2Len, 32, sizeof(int8_t));
     int32_t i = 0;
     int32_t j = 0;
     int32_t end_query = s1Len-1;
