+++ src/nw_trace_diag_avx2_256_8.c
@@ -44,7 +44,7 @@ static inline int8_t _mm256_extract_epi8_rpl(__m256i a, int imm) {
 
 
 static inline void arr_store_si256(
-        int *array,
+        int8_t *array,
         __m256i vWH,
         int32_t i,
         int32_t s1Len,
@@ -168,7 +168,7 @@ parasail_result_t* FNAME(
     int8_t * const restrict s2 = s2B+PAD; /* will allow later for negative indices */
     int8_t * const restrict H_pr = _H_pr+PAD;
     int8_t * const restrict F_pr = _F_pr+PAD;
-    parasail_result_t *result = parasail_result_new_trace(s1Len, s2Len, 32, sizeof(int));
+    parasail_result_t *result = parasail_result_new_trace(s1Len, s2Len, 32, sizeof(int8_t));
     int32_t i = 0;
     int32_t j = 0;
     int32_t end_query = s1Len-1;
